fix svg usage in examples, consolidate layout drawing, fix toggle aspect ratio
