+++ examples/buttons/src/main.rs
@@ -16,7 +16,17 @@ fn main() {
             row_spaced(
                 20.,
                 vec![
-                    toggle(id!(), binding!(State, t1)).finish(),
+                    column_spaced(
+                        20.,
+                        vec![
+                            text(id!(), "Custom Label Text")
+                                .fill(AlphaColor::WHITE)
+                                .finish(),
+                            toggle(id!(), binding!(State, t1)).finish().height(80.),
+                        ],
+                    )
+                    .height(150.)
+                    .width(200.),
                     column_spaced(
                         20.,
                         vec![
@@ -65,7 +75,7 @@ fn main() {
                             button(id!(), binding!(State, b3))
                                 .on_click(|s, _| s.count += 1)
                                 .label(|button| {
-                                    svg(id!(), "assets/download.svg")
+                                    svg(id!(), include_str!("../../../assets/download.svg"))
                                         .fill(match (button.depressed, button.hovered) {
                                             (true, _) => AlphaColor::from_rgb8(190, 190, 190),
                                             (false, true) => AlphaColor::from_rgb8(250, 250, 250),
