+++ src/editor.rs
@@ -1,10 +1,10 @@
 use backer::models::Area;
 use core::default::Default;
-use parley::{GenericFamily, StyleProperty, editor::SplitString, layout::PositionedLayoutItem};
+use parley::{GenericFamily, StyleProperty, editor::SplitString};
 use std::time::{Duration, Instant};
 use vello_svg::vello::{
     Scene,
-    kurbo::{Affine, Line, Point, RoundedRect, Stroke},
+    kurbo::{Affine, Point, RoundedRect},
     peniko::{Brush, Color, Fill, color::palette},
 };
 use winit::{event::Modifiers, keyboard::NamedKey};
@@ -12,7 +12,7 @@ use winit::{event::Modifiers, keyboard::NamedKey};
 pub use parley::layout::editor::Generation;
 use parley::{FontContext, LayoutContext, PlainEditor, PlainEditorDriver};
 
-use crate::Key;
+use crate::{Key, draw_layout::draw_layout};
 
 #[derive(Clone)]
 pub struct Editor {
@@ -340,8 +340,6 @@ impl Editor {
     ) -> Generation {
         let transform = Affine::translate((area.x as f64, area.y as f64)).then_scale(scale);
 
-        // let transform = Affine::translate((area.x as f64 * scale, area.y as f64 * scale));
-
         for (rect, _i) in self.editor.selection_geometry().iter() {
             scene.fill(
                 Fill::NonZero,
@@ -371,107 +369,7 @@ impl Editor {
         let editor = &mut self.editor;
         editor.set_width(Some(area.width));
         let layout = editor.layout(font_cx, layout_cx);
-
-        for line in layout.lines() {
-            for item in line.items() {
-                let PositionedLayoutItem::GlyphRun(glyph_run) = item else {
-                    continue;
-                };
-                let style = glyph_run.style();
-                // We draw underlines under the text, then the strikethrough on top, following:
-                // https://drafts.csswg.org/css-text-decor/#painting-order
-                if let Some(underline) = &style.underline {
-                    let underline_brush = &style.brush;
-                    let run_metrics = glyph_run.run().metrics();
-                    let offset = match underline.offset {
-                        Some(offset) => offset,
-                        None => run_metrics.underline_offset,
-                    };
-                    let width = match underline.size {
-                        Some(size) => size,
-                        None => run_metrics.underline_size,
-                    };
-                    // The `offset` is the distance from the baseline to the top of the underline
-                    // so we move the line down by half the width
-                    // Remember that we are using a y-down coordinate system
-                    // If there's a custom width, because this is an underline, we want the custom
-                    // width to go down from the default expectation
-                    let y = glyph_run.baseline() - offset + width / 2.;
-
-                    let line = Line::new(
-                        (glyph_run.offset() as f64, y as f64),
-                        ((glyph_run.offset() + glyph_run.advance()) as f64, y as f64),
-                    );
-                    scene.stroke(
-                        &Stroke::new(width.into()),
-                        transform,
-                        underline_brush,
-                        None,
-                        &line,
-                    );
-                }
-                let mut x = glyph_run.offset();
-                let y = glyph_run.baseline();
-                let run = glyph_run.run();
-                let font = run.font();
-                let font_size = run.font_size();
-                let synthesis = run.synthesis();
-                let glyph_xform = synthesis
-                    .skew()
-                    .map(|angle| Affine::skew(angle.to_radians().tan() as f64, 0.0));
-
-                scene
-                    .draw_glyphs(font)
-                    .brush(&style.brush)
-                    .hint(true)
-                    .transform(transform)
-                    .glyph_transform(glyph_xform)
-                    .font_size(font_size)
-                    .normalized_coords(run.normalized_coords())
-                    .draw(
-                        Fill::NonZero,
-                        glyph_run.glyphs().map(|glyph| {
-                            let gx = x + glyph.x;
-                            let gy = y - glyph.y;
-                            x += glyph.advance;
-                            vello_svg::vello::Glyph {
-                                id: glyph.id as _,
-                                x: gx,
-                                y: gy,
-                            }
-                        }),
-                    );
-                if let Some(strikethrough) = &style.strikethrough {
-                    let strikethrough_brush = &style.brush;
-                    let run_metrics = glyph_run.run().metrics();
-                    let offset = match strikethrough.offset {
-                        Some(offset) => offset,
-                        None => run_metrics.strikethrough_offset,
-                    };
-                    let width = match strikethrough.size {
-                        Some(size) => size,
-                        None => run_metrics.strikethrough_size,
-                    };
-                    // The `offset` is the distance from the baseline to the *top* of the strikethrough
-                    // so we calculate the middle y-position of the strikethrough based on the font's
-                    // standard strikethrough width.
-                    // Remember that we are using a y-down coordinate system
-                    let y = glyph_run.baseline() - offset + run_metrics.strikethrough_size / 2.;
-
-                    let line = Line::new(
-                        (glyph_run.offset() as f64, y as f64),
-                        ((glyph_run.offset() + glyph_run.advance()) as f64, y as f64),
-                    );
-                    scene.stroke(
-                        &Stroke::new(width.into()),
-                        transform,
-                        strikethrough_brush,
-                        None,
-                        &line,
-                    );
-                }
-            }
-        }
+        draw_layout(None, transform, layout, scene);
         self.editor.generation()
     }
 }
